--- a/nntrainer/tensor/hgemm/hgemm.h
+++ b/nntrainer/tensor/hgemm/hgemm.h
@@ -29,6 +29,22 @@ void hgemm(const __fp16 *A, const __fp16 *B, __fp16 *C, unsigned int M,
            bool TransB);
 
 /**
+ * @brief     hgemm computation with neon but with small dim without padding : Y
+ * = alpha*op(A)*op(B) + beta*C, where op(X) is one of X or X**T
+ * @param[in] A __fp16 * for Matrix A
+ * @param[in] B __fp16 * for Matrix B
+ * @param[in] C __fp16 * for Matrix C
+ * @param[in] M number of op(A)'s and C's row
+ * @param[in] N number of op(B)'s and C's columns
+ * @param[in] K number of op(A)'s and columns and op(B)'s rows
+ * @param[in] alpha float number
+ * @param[in] beta float number
+ */
+void hgemm_small(const __fp16 *A, const __fp16 *B, __fp16 *C, unsigned int M,
+                 unsigned int N, unsigned int K, float alpha, float beta,
+                 bool TransA, bool TransB);
+
+/**
  * @brief     Checking function for whether matrix A or B needs padding for
  * optimal performance of fixed blocking-kernel sequence
  * @param[in] A __fp16 * for Matrix A