--- a/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel.h
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel.h
@@ -22,21 +22,10 @@
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_8x16(unsigned int M, unsigned int N, unsigned int K,
-                       __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_8x16 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_8x16(unsigned int M, unsigned int N, unsigned int K,
-                       __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                       __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);
+
 /**
  * @brief hgemm_kernel_8x8 KERNEL function
  *
@@ -48,21 +37,10 @@ void hgemm_kernel_8x16(unsigned int M, unsigned int N, unsigned int K,
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_8x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_8x8 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_8x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                      __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);
+
 /**
  * @brief hgemm_kernel_4x8 KERNEL function
  *
@@ -74,21 +52,10 @@ void hgemm_kernel_8x8(unsigned int M, unsigned int N, unsigned int K,
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_4x8 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                      __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);
+
 /**
  * @brief hgemm_kernel_4x4 KERNEL function
  *
@@ -100,21 +67,10 @@ void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_4x4 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                      __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);
+
 /**
  * @brief hgemm_kernel_1x8 KERNEL function
  *
@@ -126,21 +82,10 @@ void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_1x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_1x8 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_1x8(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                      __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);
+
 /**
  * @brief hgemm_kernel_1x4 KERNEL function
  *
@@ -152,18 +97,6 @@ void hgemm_kernel_1x8(unsigned int M, unsigned int N, unsigned int K,
  * @param sc Starting address of blocked C
  * @param ldc Leading dimension of original matrix C
  */
+template <typename T>
 void hgemm_kernel_1x4(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc);
-/**
- * @brief hgemm_kernel_1x4 KERNEL function
- *
- * @param M Length of blocked M
- * @param N Length of blocked N
- * @param K Length of blocked K
- * @param sa Starting address of blocked A
- * @param sb Starting address of blocked B
- * @param sc Starting address of blocked C
- * @param ldc Leading dimension of original matrix C
- */
-void hgemm_kernel_1x4(unsigned int M, unsigned int N, unsigned int K,
-                      __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc);
+                      __fp16 *sa, __fp16 *sb, T *sc, unsigned int ldc);