--- a/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_1x4.cpp
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_1x4.cpp
@@ -16,6 +16,7 @@
 #include <hgemm_kernel.h>
 #include <stdlib.h>
 
+template <>
 void hgemm_kernel_1x4(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);
@@ -70,6 +71,7 @@ void hgemm_kernel_1x4(unsigned int M, unsigned int N, unsigned int K,
   }
 }
 
+template <>
 void hgemm_kernel_1x4(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);