--- a/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x4.cpp
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x4.cpp
@@ -214,6 +214,7 @@
               vaddq_f32(vld1q_f32(c + 3 * ldc), vcvt_f32_f16(v27)));      \
   } while (0)
 
+template <>
 void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);
@@ -292,6 +293,7 @@ void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
   }
 }
 
+template <>
 void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);