--- a/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x4.cpp
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x4.cpp
@@ -14,6 +14,7 @@
 #include <arm_neon.h>
 #include <assert.h>
 #include <hgemm_kernel.h>
+#include <hgemm_util.h>
 #include <stdlib.h>
 
 #define INIT_KERNEL_4x4()  \
@@ -302,8 +303,8 @@ void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
   __fp16 *a = sa, *b = sb;
   float *c = sc;
   unsigned int i, j, l;
-  unsigned int K16 = (K >> 4) << 4;
-  unsigned int K8 = (K >> 3) << 3;
+  unsigned int K8 = get_prev_mltpl_of_2p_n(K, 3);
+  unsigned int K16 = get_prev_mltpl_of_2p_n(K, 4);
   for (i = 0; i < M; i += 4) {
     for (j = 0; j < N; j += 4) {
       __builtin_prefetch(b, 0, 3);