--- a/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x8.cpp
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/hgemm_kernel_4x8.cpp
@@ -258,6 +258,7 @@
                                          vcvt_f32_f16(vget_high_f16(v9))));   \
   } while (0)
 
+template <>
 void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, __fp16 *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);
@@ -295,6 +296,7 @@ void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
   }
 }
 
+template <>
 void hgemm_kernel_4x8(unsigned int M, unsigned int N, unsigned int K,
                       __fp16 *sa, __fp16 *sb, float *sc, unsigned int ldc) {
   assert(M > 0 && N > 0 && K > 0);