--- a/nntrainer/tensor/hgemm/hgemm_kernel/meson.build
+++ b/nntrainer/tensor/hgemm/hgemm_kernel/meson.build
@@ -9,8 +9,13 @@ hgemm_kernel_sources = [
     'hgemm_kernel_4x4.cpp',
     'hgemm_kernel_4x8.cpp',
     'hgemm_kernel_8x8.cpp',
-    'hgemm_kernel_8x16.cpp',
 ]
+
+if get_option('hgemm-experimental-kernel')
+  hgemm_kernel_sources += 'hgemm_kernel_8x16_experimental.cpp'
+else
+  hgemm_kernel_sources += 'hgemm_kernel_8x16.cpp'
+endif
 
 foreach s : hgemm_kernel_sources
   nntrainer_sources += meson.current_source_dir() / s