--- a/nntrainer/tensor/hgemm/hgemm_noTrans.h
+++ b/nntrainer/tensor/hgemm/hgemm_noTrans.h
@@ -252,6 +252,28 @@ void hgemm_noTrans_8x16(unsigned int M, unsigned int N, unsigned int K,
                         float alpha = 1.F, float beta = 0.F);
 
 /**
+ * @brief hgemm noTrans computation with 8x16 kernel : C = A*B,
+ *
+ * @param M length of the row of matrix A
+ * @param N length of the col of matrix B
+ * @param K length of the col of matrix A
+ * @param A input matrix A
+ * @param lda length of the col of matrix A
+ * @param B input matrix B
+ * @param ldb length of the col of matrix B
+ * @param C output matrix C
+ * @param ldc length of the col of matrix C
+ * @param[in] alpha float number
+ * @param[in] beta float number
+ */
+void hgemm_noTrans_8x16_experimental(unsigned int M, unsigned int N,
+                                     unsigned int K, const __fp16 *A,
+                                     unsigned int lda, const __fp16 *B,
+                                     unsigned int ldb, float *C,
+                                     unsigned int ldc, float alpha = 1.F,
+                                     float beta = 0.F);
+
+/**
  * @brief     hgemm fallback with NEON : Y = alpha*op(A)*op(B) + beta*C,
  * @param M length of the row of matrix A
  * @param N length of the col of matrix B