- According to current paper, accumulating up to 64 ~ 128 w.r.t. K-direction is fine.
- Since conventional error metric, and newly introduced metric (max component relative error) is fine as well, introduce experiemntal kernel that is faster, but less accurate.
- using build option `-Dhgemm-experimental-kernel=true` can enable such kernel when *android build*.
- Performance of this kernel is shown at  #2655 
- Include experimental kernel and exclude the previous one when experimental kernel build -> no newly added `#ifdef`s in file