@skykongkong8 2024-08-07
> Overall, awesome work! One thing I want to point out is the newly added macros.
> 
> 1. There's room for the code to be reduced.
> 2. I think four macros can be combined into one.
> 
> How about replacing it as follows?
> 
> ```
> // N would be 1, 4, 8, and 16
> #define KERNEL_8x16_ACC(N)                               \
>   do {                                                   \
>     for (int i = 0; i < N; ++i) {                        \
>       va0 = vld1q_f16(a + 8 * i);                        \
>       vb1 = vld1q_f16(b + 16 * i);                       \
>       vb2 = vld1q_f16(b + 16 * i + 8);                   \
>       v0_7 = vfmaq_laneq_f16(v0_7, vb1, va0, 0);         \
>       v8_15 = vfmaq_laneq_f16(v8_15, vb1, va0, 1);       \
>       v16_23 = vfmaq_laneq_f16(v16_23, vb1, va0, 2);     \
>       v24_31 = vfmaq_laneq_f16(v24_31, vb1, va0, 3);     \
>       v32_39 = vfmaq_laneq_f16(v32_39, vb1, va0, 4);     \
>       v40_47 = vfmaq_laneq_f16(v40_47, vb1, va0, 5);     \
>       v48_55 = vfmaq_laneq_f16(v48_55, vb1, va0, 6);     \
>       v56_63 = vfmaq_laneq_f16(v56_63, vb1, va0, 7);     \
>       v64_71 = vfmaq_laneq_f16(v64_71, vb2, va0, 0);     \
>       v72_79 = vfmaq_laneq_f16(v72_79, vb2, va0, 1);     \
>       v80_87 = vfmaq_laneq_f16(v80_87, vb2, va0, 2);     \
>       v88_95 = vfmaq_laneq_f16(v88_95, vb2, va0, 3);     \
>       v96_103 = vfmaq_laneq_f16(v96_103, vb2, va0, 4);   \
>       v104_111 = vfmaq_laneq_f16(v104_111, vb2, va0, 5); \
>       v112_119 = vfmaq_laneq_f16(v112_119, vb2, va0, 6); \
>       v120_127 = vfmaq_laneq_f16(v120_127, vb2, va0, 7); \
>     }                                                    \
>                                                          \
>     l += N;                                              \
>     __builtin_prefetch(b + 16 * N, 0, 3);                \
>     __builtin_prefetch(a + 8 * N, 0, 3);                 \
>     b += 16 * N;                                         \
>     a += 8 * N;                                          \
>   } while (0);
> ```
> 
> If this works, it can reduce about 600 lines of code. would this be valid?

This would be invalid according to my research. For some reason, many kernels are designed up to write hard-coded lines rather than looping through them. Some says it would make the kernel efficiency differ from time to time. So I just coded this way.
Observing GEMM kernels in XNNPACK, or OpenBLAS would help you with this.