@skykongkong8 2024-08-07
@djeong20 
Hmm... I've gone through FP16 GEMM unittest test, and had a little bit unhappy result.
In short, they have very subtle but clear difference...
smallest difference in terms of Glops is 1%, but at biggest is 8%
Following is TC result for your information

```bash
Before
Try #1
[INFO] Latency : 4095.28490000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 93.32829372
[INFO] Latency : 6330.02291000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 95.88099516
[INFO] Latency : 9644.96927000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 93.93183520
[INFO] Latency : 14042.25833000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 91.86165485
[INFO] Latency : 18921.98801000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 93.51406412
[INFO] Latency : 26311.99010000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 89.50927783

Try #2
[INFO] Latency : 4032.19531000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 94.78855130
[INFO] Latency : 6488.52344000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 93.53883077
[INFO] Latency : 9809.80469000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 92.35348640
[INFO] Latency : 14213.26614000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 90.75641554
[INFO] Latency : 19063.69687000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 92.81893287
[INFO] Latency : 26564.81249000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 88.65740095

Try #3
[INFO] Latency : 4079.07395000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 93.69919660
[INFO] Latency : 6545.82968000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 92.71993401
[INFO] Latency : 9888.91458000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 91.61467183
[INFO] Latency : 14376.37395000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 89.72673447
[INFO] Latency : 19377.54166000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 91.31560809
[INFO] Latency : 26971.74739000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 87.31978681

Try #4
[INFO] Latency : 4142.87708000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 92.25616513
[INFO] Latency : 6629.20677000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 91.55377363
[INFO] Latency : 10061.58437000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 90.04244567
[INFO] Latency : 14656.08801000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 88.01428370
[INFO] Latency : 19676.60885000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 89.92769097
[INFO] Latency : 27258.33019000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 86.40174272

After
Try #1
[INFO] Latency : 4141.33907000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 92.29042721
[INFO] Latency : 6470.47916000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 93.79968330
[INFO] Latency : 9858.99531000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 91.89269652
[INFO] Latency : 14423.61198000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 89.43287505
[INFO] Latency : 19305.43697000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 91.65666660
[INFO] Latency : 26863.52916000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 87.67154970

Try #2
[INFO] Latency : 4083.25104000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 93.60334406
[INFO] Latency : 6561.86666000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 92.49332963
[INFO] Latency : 10031.08906000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 90.31618188
[INFO] Latency : 14521.63593000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 88.82918524
[INFO] Latency : 19558.44113000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 90.47101393
[INFO] Latency : 27245.39114000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 86.44277558

Try#3
[INFO] Latency : 4223.33125000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 90.49869152
[INFO] Latency : 6735.30781000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 90.11153063
[INFO] Latency : 10125.71198000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 89.47219374
[INFO] Latency : 14758.81614000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 87.40166391
[INFO] Latency : 19903.89791000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 88.90077753
[INFO] Latency : 27369.10885000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 86.05202474


Try#4
[INFO] Latency : 4289.84010000 ms | Dim: 576 | Error: 0.00140162 | Gflops: 89.09561734
[INFO] Latency : 6892.05364000 ms | Dim: 672 | Error: 0.00197730 | Gflops: 88.06212599
[INFO] Latency : 10414.10313000 ms | Dim: 768 | Error: 0.00310214 | Gflops: 86.99449705
[INFO] Latency : 14986.09218000 ms | Dim: 864 | Error: 0.00367473 | Gflops: 86.07614797
[INFO] Latency : 20215.67447000 ms | Dim: 960 | Error: 0.01159486 | Gflops: 87.52970388
[INFO] Latency : 27385.74374000 ms | Dim: 1056 | Error: 0.00553330 | Gflops: 85.99975426
```